+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.10.12] - 2022-08-04
 
+### Added
+- Add holiday exceptions
+
 ### Fixed
 
 - Destructuring error in vbase `alreadyHasSitemap` check